// src/components/OutsideClickDetector/OutsideClickDetector.types.ts
// Generated with util/create-component.js
import { IProps } from "../../interfaces/IProps";

export interface OutsideClickDetectorProps extends IProps {
    onOutsideClick: (param: any) => any;
    onMouseUp?: (param: any) => any;
    onMouseDown?: (param: any) => any;
    onTouchStart?: (param: any) => any;
    onTouchEnd?: (param: any) => any;
    children?: any;
    isDisabled?: boolean;
}